+++ src/dockerHighlight.ts
@@ -8,6 +8,7 @@ import {
 	TextDocument, Position, DocumentHighlight, DocumentHighlightKind
 } from 'vscode-languageserver';
 import { DockerfileParser } from './parser/dockerfileParser';
+import { From } from './parser/instructions/from';
 import { DockerDefinition } from './dockerDefinition';
 import { Util } from './docker';
 
@@ -18,7 +19,7 @@ export class DockerHighlight {
 		let dockerfile = parser.parse(document);
 		let provider = new DockerDefinition();
 		let location = provider.computeDefinition(document, position);
-		let image = dockerfile.getContainingImage(position);
+		let image = location === null ? dockerfile.getContainingImage(position) : dockerfile.getContainingImage(location.range.start);
 		let highlights = [];
 		if (location === null) {
 			for (let instruction of dockerfile.getCOPYs()) {
@@ -36,15 +37,33 @@ export class DockerHighlight {
 				}
 			}
 
+			for (const from of dockerfile.getFROMs()) {
+				for (const variable of from.getVariables()) {
+					if (Util.isInsideRange(position, variable.getNameRange())) {
+						const name = variable.getName();
+						for (const loopFrom of dockerfile.getFROMs()) {
+							for (const fromVariable of loopFrom.getVariables()) {
+								if (fromVariable.getName() === name) {
+									highlights.push(DocumentHighlight.create(fromVariable.getNameRange(), DocumentHighlightKind.Read));
+								}
+							}
+						}
+						return highlights;
+					}
+				}
+			}
+
 			for (let instruction of image.getInstructions()) {
 				for (let variable of instruction.getVariables()) {
 					if (Util.isInsideRange(position, variable.getNameRange())) {
 						let name = variable.getName();
 						
 						for (let instruction of image.getInstructions()) {
-							for (let variable of instruction.getVariables()) {
-								if (variable.getName() === name) {
-									highlights.push(DocumentHighlight.create(variable.getNameRange(), DocumentHighlightKind.Read));
+							if (!(instruction instanceof From)) {
+								for (let variable of instruction.getVariables()) {
+									if (variable.getName() === name) {
+										highlights.push(DocumentHighlight.create(variable.getNameRange(), DocumentHighlightKind.Read));
+									}
 								}
 							}
 						}
@@ -86,9 +105,25 @@ export class DockerHighlight {
 			}
 
 			for (let instruction of image.getInstructions()) {
-				for (let variable of instruction.getVariables()) {
-					if (variable.getName() === definition) {
-						highlights.push(DocumentHighlight.create(variable.getNameRange(), DocumentHighlightKind.Read));
+				// only highlight variables in non-FROM instructions
+				if (!(instruction instanceof From)) {
+					for (const variable of instruction.getVariables()) {
+						if (variable.getName() === definition) {
+							highlights.push(DocumentHighlight.create(variable.getNameRange(), DocumentHighlightKind.Read));
+						}
+					}
+				}
+			}
+
+			for (const arg of dockerfile.getInitialARGs()) {
+				const property = arg.getProperty();
+				if (property && Util.rangeEquals(property.getNameRange(), location.range)) {
+					for (const from of dockerfile.getFROMs()) {
+						for (const variable of from.getVariables()) {
+							if (variable.getName() === definition) {
+								highlights.push(DocumentHighlight.create(variable.getNameRange(), DocumentHighlightKind.Read));
+							}
+						}
 					}
 				}
 			}